requirements.txt
deepgram-sdk>=3.0.0,<4.0.0
pyaudio>=0.2.14
python-dotenv>=1.0.0
requests>=2.31.0
pynput>=1.7.6
pyautogui>=0.9.54
pyperclip>=1.8.0
pystray>=0.19.5
Pillow>=10.0.0
